removing type file
