+++ packages/react/src/lib/useMicrophone/index.ts
@@ -1,4 +1,3 @@
 export * from './useMicrophone';
 export * from './constants';
-export * from './types';
 export * from './useEncoding';
